Browser tests: reduce timeout for instantiation to reasonable value
The timeout for browser instantiation in the browser tests has been
increased to avoid flaky tests. However, browser instantiation taking
that long is an indicator for something being wrong with the
instantiation. There is no reason why instantiation may take that long
during test execution but the same issue should never arise in
productive use, thus such a case should be captured by tests and, if
necessary, be fixed productively.

In order to do so, this change reduces the test timeout value again to a
more reasonable value. Since the first Edge browser instantiation takes
rather long at least in the GitHub actions test environment, that
instantiation is isolated in an execution before the test class.