Unify and improve image disablement calculation algorithm
Each OS-specific implementation of the Image class has its own
implementation of an algorithm to calculate a disabled representation of
a given image. In addition, the algorithm used on Windows and MacOS
produces bad results, which is why usually pre-generated disabled icons
are used.

This change extracts the implementations of those algorithms into
ImageProcessingDescriptors. The implementation for Windows and MacOS is
improved and a legacy mode applying the old algorithm is added. The
extraction of the processing also allows to later on add a
(pre-)processing implementation for SVG images