[Win32] Remove OS version retrieval from OS class
The OS version retrieval has been moved to the separate class/DLL
OsVersion to be used for OS version validation at application startup.
The methods previously provided via the OS class are thus obsolete and
can be removed. This change removes the methods that became obsolete and
moves everything related to OS versions to the OsVersion class.