[Win32] Avoid broken UI when setting disposed image on rescaling #2432
In case an image set to an Item has already been disposed when a DPI
change event occurs, a DPI change handler may try to set that image on
the image again, leading to an exception because of the image being
disposed.

This change ensures that such an exception happening when setting a
(potentially disposed) iamge does not result in a complete rescaling
failure and instead reports the listener error to the Display.

Contributes to
https://github.com/eclipse-platform/eclipse.platform.swt/issues/2432