[Win32] Autoscale cursor even on widgets that have autoscaling disabled
Cursors are initialized with the zoom of the control on which they are
set. However, cursors are usually expected to scale according to the
autoscaled UI around those controls for which autoscaling was disabled.

This change adapts the cursor initializations to use the autoscale zoom
of the containing shell instead of the zoom of the control itself. In
consequence, cursors will have the same size in the whole shell, no
matter on which control they are set.